voice-deepfake-detector/requirements.txt
torch
torchaudio
transformers
joblib
datasets
numpy
pandas
scikit-learn
tqdm
matplotlib
pyyaml
soundfile
sounddevice
